refactor: move nineties popup methods to utils
